refactor(p2p) - connection_slot modularized
