+++ hathor/p2p/peer_discovery/dns.py
@@ -59,7 +59,8 @@ class DNSPeerDiscovery(PeerDiscovery):
         """
         for host in self.hosts:
             for entrypoint in (await self.dns_seed_lookup(host)):
-                connect_to_endpoint(entrypoint, discovery_call=True)  # type: ignore[call-arg]
+                connect_to_endpoint(entrypoint, discovery_call=True)
+
 
     async def dns_seed_lookup(self, host: str) -> set[PeerEndpoint]:
         """ Run a DNS lookup for TXT, A, and AAAA records and return a list of connection strings.
